782: move the capistrano2 logic back to rollbar/capistrano.rb
