+++ lib/rollbar/tasks/rollbar.cap
@@ -6,8 +6,6 @@ require 'byebug'
 
 namespace :rollbar do
 
-  # TODO: adjust code for capistrano 2
-
   desc 'Send deployment started notification to Rollbar.'
   task :deploy_started do
     on primary fetch(:rollbar_role) do
